product-carrot/requirements.txt
Django==4.2.3
django-bootstrap4==23.2
django-stdimage==6.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.6
djangorestframework-simplejwt==5.2.2
